requirements.txt
pandas>=0.23
requests>=2.22.0
rapidfuzz>=0.2.2
ruamel.yaml>=0.16.10
click>=7.0
inquirer>=2.6.3
colorama>=0.4.3
GitPython>=3.1.0
loguru>=0.5.3
datapackage>=1.15.2
awscli==1.20.13
docutils<0.16,>=0.10
mkdocs-material==7.1.8
python-slugify==5.0.2
mkdocs-macros-plugin==0.5.12
rich>=10.7.0
pytest==7.1.2
validators==0.19.0
yarl==1.7.2
giturlparse==0.10.0
